+++ src/dftbp/md/thermostats.F90
@@ -16,6 +16,7 @@ module dftbp_md_thermostats
   use dftbp_md_andersentherm, only : TAndersenTherm, TAndersenTherm_init, TAndersenThermInput
   use dftbp_md_berendsentherm, only : TBerendsenTherm, TBerendsenTherm_init, TBerendsenThermInput
   use dftbp_md_notherm, only : TNoTherm, TNoTherm_init
+  use dftbp_md_langevintherm, only : TLangevinTherm, TLangevinTherm_init, TLangevinThermInput
   use dftbp_md_nhctherm, only : TNhcTherm, TNhcTherm_init, TNhcThermInput
   use dftbp_md_thermostat, only : TThermostat
   implicit none
@@ -25,6 +26,7 @@ module dftbp_md_thermostats
   public :: createThermostat, TThermostat, TThermostatInput
   public :: TAndersenTherm, TAndersenTherm_init, TAndersenThermInput
   public :: TBerendsenTherm, TBerendsenTherm_init, TBerendsenThermInput
+  public :: TLangevinTherm, TLangevinTherm_init, TLangevinThermInput
   public :: TNoTherm, TNoTherm_init
   public :: TNhcTherm, TNhcTherm_init, TNhcThermInput
 
@@ -34,6 +36,7 @@ module dftbp_md_thermostats
     integer :: andersen = 1
     integer :: berendsen = 2
     integer :: nhc = 3
+    integer :: langevin = 4
   end type TThermostatTypes_
 
   !> Available thermostat types
@@ -45,6 +48,7 @@ module dftbp_md_thermostats
     integer :: thermostatType
     type(TAndersenThermInput), allocatable :: andersen
     type(TBerendsenThermInput), allocatable :: berendsen
+    type(TLangevinThermInput), allocatable :: langevin
     type(TNhcThermInput), allocatable :: nhc
   end type TThermostatInput
 
@@ -79,15 +83,19 @@ contains
     type(TNoTherm), allocatable :: noTherm
     type(TAndersenTherm), allocatable :: andersenTherm
     type(TBerendsenTherm), allocatable :: berendsenTherm
+    type(TLangevinTherm), allocatable :: langevinTherm
     type(TNHCTherm), allocatable :: nhcTherm
     real(dp) :: tempAtom
 
-    @:ASSERT(input%thermostatType >= 0 .and. input%thermostatType <= 3)
+    @:ASSERT(input%thermostatType >= thermostatTypes%none .and.&
+        & input%thermostatType <= thermostatTypes%langevin)
     @:ASSERT(all(&
-        & ([thermostatTypes%andersen, thermostatTypes%berendsen, thermostatTypes%nhc]&
+        & ([thermostatTypes%andersen, thermostatTypes%berendsen, thermostatTypes%nhc,&
+        & thermostatTypes%langevin]&
         & == input%thermostatType)&
         & .eqv. &
-        & [allocated(input%andersen), allocated(input%berendsen), allocated(input%nhc)]&
+        & [allocated(input%andersen), allocated(input%berendsen), allocated(input%nhc),&
+        & allocated(input%langevin)]&
         & ))
 
     select case (input%thermostatType)
@@ -106,6 +114,11 @@ contains
       call TBerendsenTherm_init(berendsenTherm, input%berendsen, randomThermostat, masses,&
           ptempProfile, pMdFrame)
       call move_alloc(berendsenTherm, thermostat)
+    case (thermostatTypes%langevin)
+      allocate(langevinTherm)
+      call TLangevinTherm_init(langevinTherm, input%langevin, randomThermostat, masses,&
+          ptempProfile, pMdFrame, deltaT)
+      call move_alloc(langevinTherm, thermostat)
     case (thermostatTypes%nhc)
       allocate(nhcTherm)
       call TNhcTherm_init(nhcTherm, input%nhc, randomThermostat, masses, ptempProfile,&